Expand unit test coverage for core and infra components
Added comprehensive unit tests for Xtate core, persistence, state machine host, resource loader, and SCXML infrastructure. Introduced new coverage tests for converters, storage, controllers, service execution, resource handlers, and validators. Improved coverage for dynamic behaviors, error handling, and edge cases. Fixed and clarified test assertions, updated test data and XML samples, and marked known defects. Re-enabled two regression tests and performed minor code cleanups. Updated coverage tracker and ensured all changes are AGPL-compliant.

+++ src/Xtate.Core/Persistence/Services/Bucket.cs
@@ -572,7 +572,7 @@ public readonly struct Bucket
 
 		protected override int Get(ReadOnlySpan<byte> bytes)
 		{
-			var value = (int) bytes[^1];
+			var value = (int) (sbyte) bytes[^1];
 
 			for (var i = bytes.Length - 2; i >= 0; i --)
 			{
